packages/script-generator/pyproject.toml
[project]
name = "script-generator"
version = "0.1.0"
description = "Debate script generation using Gemini API"
requires-python = ">=3.12"
dependencies = [
    "shared",
    "google-generativeai>=0.7",
    "pydantic>=2.7",
    "jinja2>=3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/script_generator"]
